Moves transform file into transform module
